feat. 商品详情页及数据控制开发

+++ datamodels/order.go
@@ -4,7 +4,7 @@ type Order struct {
 	Id          int64 `sql:"id"`
 	UserId      int64 `sql:"userId"`
 	ProductId   int64 `sql:"productId"`
-	OrderStatus int64 `sql:"orderStatus"`
+	OrderStatus int   `sql:"orderStatus"`
 }
 
 const (
